requirements.txt
# local package install
-e .

# third party packages
seaborn
matplotlib
pandas~=1.5.3
flask~=2.2.2
pytest
jupyter
PyYAML~=6.0
numpy~=1.23.5
sklearn~=0.0.post1
scikit-learn~=1.2.1
joblib~=1.2
setuptools~=60.2.0
